delete useless scripts and fix some errors

+++ engine/Scripts/lib_build.py
@@ -286,7 +286,8 @@ def build_engine():
                 tundra_file=Path(work_path + "/../artifacts/tundra.dag.json")
                 prelinkfiles(tundra_file, runtime_mode, output_path, work_path, bitcode)
     elif runtime_mode == "debug":
-        shutil.rmtree(Path(work_path + "/../build_debug"))
+        if os.path.exists(Path(work_path + "/../build_debug")):
+            shutil.rmtree(Path(work_path + "/../build_debug"))
         if platform == "windows":
             os.system("bee.exe win_debug")
         else:
